feat(portal): inline ReportView section + per-section nav infra (Hub + standalone + agent reuse)
